refactor: 라우트 세그먼트에 따라 레이아웃 분기되도록 구조 수정
