+++ app/admin/(dashboard)/course-info/page.tsx
