+++ app/admin/(dashboard)/layout.tsx
@@ -1,0 +1,25 @@
+import type React from "react";
+import type { Metadata } from "next";
+import { Inter } from "next/font/google";
+import Sidebar from "@/components/admin/sidebar";
+
+const inter = Inter({ subsets: ["latin"] });
+
+export const metadata: Metadata = {
+  title: "demo-funnel Admin",
+  description: "Admin for demo-funnel",
+  generator: "v0.dev",
+};
+
+export default function DashboardLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <div className="flex h-screen">
+      <Sidebar />
+      <main className="flex-1 overflow-y-auto p-8">{children}</main>
+    </div>
+  );
+}
