+++ app/admin/(dashboard)/lectures/page.tsx
